back-end/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
Automat==22.10.0
Babel==2.14.0
bleach==6.1.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.16.1
Flask==3.1.0
fqdn==1.5.1
fsspec==2024.10.0
hyperlink==21.0.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
lxml==5.2.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.1.3
notebook_shim==0.2.4
numpy==2.1.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pillow==11.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.3.1
psutil==5.9.8
pure-eval==0.2.2
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
requests-file==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Scrapy==2.11.1
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==75.1.0
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
tinycss2==1.3.0
tldextract==5.1.2
torch==2.5.1
torchvision==0.20.1
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uri-template==1.3.0
w3lib==2.1.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.10
zope.interface==6.3
